Safari-611.1.10/LayoutTests/svg/animations/svginteger-animation-2-expected.txt

SVG 1.1 dynamic animation tests

Test animation of SVGInteger.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feConvolveMatrix.orderX.animVal is 3
PASS feConvolveMatrix.orderY.animVal is 1
PASS feConvolveMatrix.orderX.baseVal is 3
PASS feConvolveMatrix.orderY.baseVal is 1
PASS feConvolveMatrix.orderX.animVal is 2
PASS feConvolveMatrix.orderY.animVal is 2
PASS feConvolveMatrix.orderX.baseVal is 3
PASS feConvolveMatrix.orderY.baseVal is 1
PASS feConvolveMatrix.orderX.animVal is 1
PASS feConvolveMatrix.orderY.animVal is 3
PASS feConvolveMatrix.orderX.baseVal is 3
PASS feConvolveMatrix.orderY.baseVal is 1
PASS feConvolveMatrix.orderX.animVal is 3
PASS feConvolveMatrix.orderY.animVal is 1
PASS feConvolveMatrix.orderX.baseVal is 3
PASS feConvolveMatrix.orderY.baseVal is 1
PASS successfullyParsed is true

TEST COMPLETE
